// src/Utilities/Enums/Tiles.ts
export const Flags = {
    FLAGS_TILEEXTRA: 0x0001,
    FLAGS_LOCKED: 0x0002,
    FLAGS_SEED: 0x0010,
    FLAGS_TREE: 0x0019,
    FLAGS_FLIPPED: 0x0020,
    FLAGS_ROTATED_LEFT: 0x0030,
    FLAGS_OPEN: 0x0040,
    FLAGS_PUBLIC: 0x0080,
    FLAGS_SILENCED: 0x0200,
    FLAGS_WATER: 0x0400,
    FLAGS_FIRE: 0x1000,
    FLAGS_RED: 0x2000,
    FLAGS_BLUE: 0x4000,
    FLAGS_GREEN: 0x8000
  };
  
  export enum ActionTypes {
    FIST = 0,
    WRENCH = 1,
    DOOR = 2,
    LOCK = 3,
    GEMS = 4,
    TREASURE = 5,
    DEADLY_BLOCK = 6,
    TRAMPOLINE = 7,
    CONSUMABLE = 8,
    GATEWAY = 9,
    SIGN = 10,
    SFX_WITH_EXTRA_FRAME = 11,
    BOOMBOX = 12,
    MAIN_DOOR = 13,
    PLATFORM = 14,
    BEDROCK = 15,
    LAVA = 16,
    FOREGROUND = 17,
    BACKGROUND = 18,
    SEED = 19,
    CLOTHES = 20,
    FOREGROUND_WITH_EXTRA_FRAME = 21,
    BACKGD_SFX_EXTRA_FRAME = 22,
    BACK_BOOMBOX = 23,
    BOUNCY = 24,
    POINTY = 25,
    PORTAL = 26,
    CHECKPOINT = 27,
    SHEET_MUSIC = 28,
    ICE = 29,
    SWITCHEROO = 31,
    CHEST = 32,
    MAILBOX = 33,
    BULLETIN = 34,
    PINATA = 35,
    DICE = 36,
    CHEMICAL = 37,
    PROVIDER = 38,
    LAB = 39,
    ACHIEVEMENT = 40,
    WEATHER_MACHINE = 41,
    SCORE_BOARD = 42,
    SUNGATE = 43,
    PROFILE = 44,
    DEADLY_IF_ON = 45,
    HEART_MONITOR = 46,
    DONATION_BOX = 47,
    TOYBOX = 48,
    MANNEQUIN = 49,
    SECURITY_CAMERA = 50,
    MAGIC_EGG = 51,
    GAME_RESOURCES = 52,
    GAME_GENERATOR = 53,
    XENONITE = 54,
    DRESSUP = 55,
    CRYSTAL = 56,
    BURGLAR = 57,
    COMPACTOR = 58,
    SPOTLIGHT = 59,
    WIND = 60,
    DISPLAY_BLOCK = 61,
    VENDING_MACHINE = 62,
    FISHTANK = 63,
    PETFISH = 64,
    SOLAR = 65,
    FORGE = 66,
    GIVING_TREE = 67,
    GIVING_TREE_STUMP = 68,
    STEAMPUNK = 69,
    STEAM_LAVA_IF_ON = 70,
    STEAM_ORGAN = 71,
    TAMAGOTCHI = 72,
    SWING = 73,
    FLAG = 74,
    LOBSTER_TRAP = 75,
    ART_CANVAS = 76,
    BATTLE_CAGE = 77,
    PET_TRAINER = 78,
    STEAM_ENGINE = 79,
    LOCKBOT = 80,
    WEATHER_SPECIAL = 81,
    SPIRIT_STORAGE = 82,
    DISPLAY_SHELF = 83,
    VIP_ENTRANCE = 84,
    CHALLENGE_TIMER = 85,
    CHALLENGE_FLAG = 86,
    FISH_MOUNT = 87,
    PORTRAIT = 88,
    WEATHER_SPECIAL2 = 89,
    FOSSIL = 90,
    FOSSIL_PREP = 91,
    DNA_MACHINE = 92,
    BLASTER = 93,
    VALHOWLA = 94,
    CHEMSYNTH = 95,
    CHEMTANK = 96,
    STORAGE = 97,
    OVEN = 98,
    SUPER_MUSIC = 99,
    GEIGER_CHARGER = 100,
    ADVENTURE_RESET = 101,
    TOMB_ROBBER = 102,
    FACTION = 103,
    RED_FACTION = 104,
    GREEN_FACTION = 105,
    BLUE_FACTION = 106,
    ANCES = 107,
    FISHGOTCHI_TANK = 109,
    FISHING_BLOCK = 110,
    ITEM_SUCKER = 111,
    ITEM_PLANTER = 112,
    ROBOT = 113,
    COMMAND = 114,
    TICKET = 115,
    STATS_BLOCK = 116,
    FIELD_NODE = 117,
    OUIJA_BOARD = 118,
    ARCHITECT_MACHINE = 119,
    STARSHIP = 120,
    AUTODELETE = 121,
    GREEN_FOUNTAIN = 122,
    AUTO_ACTION_BREAK = 123,
    AUTO_ACTION_HARVEST = 124,
    AUTO_ACTION_HARVEST_SUCK = 125,
    LIGHTNING_IF_ON = 126,
    PHASED_BLOCK = 127,
    MUD = 128,
    ROOT_CUTTING = 129,
    PASSWORD_STORAGE = 130,
    PHASED_BLOCK_2 = 131,
    BOMB = 132,
    WEATHER_INFINITY = 134,
    SLIME = 135,
    UNK1 = 136,
    COMPLETIONIST = 137,
    UNK3 = 138,
    FEEDING_BLOCK = 140,
    KRANKENS_BLOCK = 141,
    FRIENDS_ENTRANCE = 142
  }
  
  export const Options = {
    MUSIC_BLOCKS_DISABLED: 0x10,
    MUSIC_BLOCKS_INVIS: 0x20
  };
  
  export enum ExtraTypes {
    DOOR = 0x01,
    MAIN_DOOR = 0x01,
    SIGN = 0x02,
    LOCK = 0x03,
    TREE = 0x04,
    MAILBOX = 0x005,
    BULLETIN_BOARD = 0x06,
    DICE = 0x07,
    PROVIDER = 0x09,
    HEART_MONITOR = 0x0b,
    ACHIEVEMENT_BLOCK = 0x10,
    DONATION_BOX = 0x12,
    MANNEQUIN = 0x14,
    BUNNY_EGG = 0x15,
    GAME_RESOURCES = 0x16,
    DISPLAY_BLOCK = 0x17,
    VENDING_MACHINE = 0x18,
    // GAME_GENERATOR = 0x17,
    // XENONITE_CRYSTAL = 0x18,
    PHONE_BOOTH = 0x19,
    CRYSTAL = 0x20
  }